+++ package.json
@@ -11,6 +11,9 @@
     },
     "devDependencies": {
         "@babel/preset-react": "^7.0.0",
+        "@types/node": "^14.14.2",
+        "@types/react": "^16.9.53",
+        "@types/react-dom": "^16.9.8",
         "axios": "^0.19",
         "babel-eslint": "^10.1.0",
         "bootstrap": "^4.0.0",
@@ -36,10 +39,13 @@
         "redux": "^4.0.5",
         "resolve-url-loader": "^3.1.0",
         "sass": "^1.15.2",
-        "sass-loader": "^8.0.0"
+        "sass-loader": "^8.0.0",
+        "ts-loader": "^8.0.6",
+        "typescript": "^4.0.3"
     },
     "dependencies": {
         "@emotion/core": "^10.0.35",
+        "@types/react-router-dom": "^5.1.6",
         "react-router-dom": "^5.2.0"
     }
 }
