--- resources/js/Style/index.js
@@ -1,8 +1,0 @@
-export default {
-    color: {
-        main: "#8CD790",
-        accent: "#77AF9C",
-        bg: "#D9EDC1",
-        lightBg: "#C5E99B",
-    },
-};
