--- resources/js/actions/index.js
@@ -1,6 +1,0 @@
-export const REGISTER = "REGISTER";
-export const LOGIN = "LOGIN";
-export const LOGOUT = "LOGOUT";
-export const AUTHCHECK = "AUTHCHECK";
-
-export const LOAD = "LOAD";
