+++ resources/js/actions/index.ts
@@ -1,0 +1,6 @@
+export const REGISTER = "REGISTER";
+export const LOGIN = "LOGIN";
+export const LOGOUT = "LOGOUT";
+export const AUTHCHECK = "AUTHCHECK";
+
+export const LOAD = "LOAD";
