--- resources/js/components/Layout/Layout.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Header from "../Header/Header";
-
-const Layout = ({ children }) => {
-    return (
-        <React.Fragment>
-            <Header />
-            {children}
-        </React.Fragment>
-    );
-};
-
-export default Layout;
