+++ resources/js/components/Layout/Layout.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import Header from "../Header/Header";
+
+const Layout = ({ children }) => {
+    return (
+        <React.Fragment>
+            <Header />
+            {children}
+        </React.Fragment>
+    );
+};
+
+export default Layout;
