--- resources/js/pages/Sample.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Sample = () => {
-    return <div>ログインしたときだけ表示</div>;
-};
-
-export default Sample;
