+++ resources/js/pages/Sample.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Sample = () => {
+    return <div>ログインしたときだけ表示</div>;
+};
+
+export default Sample;
