--- resources/js/reducers/index.js
@@ -1,6 +1,0 @@
-import { combineReducers } from "redux";
-import auth from "./auth";
-
-export default combineReducers({
-    auth,
-});
