+++ resources/js/reducers/index.ts
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import auth from "./auth";
+
+export default combineReducers({
+    auth,
+});
