+++ resources/js/util.ts
@@ -1,0 +1,21 @@
+/**
+ * クッキーの値を取得する
+ * @param {String} searchKey 検索するキー
+ * @returns {String} キーに対応する値
+ */
+export function getCookieValue(searchKey: string) {
+    if (typeof searchKey === "undefined") {
+        return "";
+    }
+
+    let val = "";
+
+    document.cookie.split(";").forEach((cookie) => {
+        const [key, value] = cookie.split("=");
+        if (key === searchKey) {
+            return (val = value);
+        }
+    });
+    //   console.log("val", val)
+    return val;
+}
